Refine SQLite backfill migration runtime handling
