+++ tui/ANATOMY.md
@@ -29,7 +29,7 @@ This folder is the self-contained Go module for the `lingtai-tui` terminal UI bi
 ## Connections
 
 - **Called from:** the shell (`lingtai-tui`), Homebrew tap (`lingtai-ai/lingtai/lingtai-tui`), `install.sh`.
-- **Calls out:** `tui/internal/tui` (Bubble Tea app), `tui/internal/migrate` (per-project migrations), `tui/internal/globalmigrate` (per-machine migrations), `tui/internal/preset` (bootstrap + utility skill population), `tui/internal/process` (agent launch), `tui/internal/config` (global config, venv, upgrade checks), `tui/internal/postman` (mail relay daemon), `tui/internal/timemachine` (git history daemon).
+- **Calls out:** `tui/internal/tui` (Bubble Tea app), `tui/internal/migrate` (per-project migrations), `tui/internal/globalmigrate` (per-machine migrations), `tui/internal/preset` (bootstrap + utility skill population), `tui/internal/process` (agent launch), `tui/internal/processscan` (shared ps-based agent-process detection), `tui/internal/config` (global config, venv, upgrade checks), `tui/internal/postman` (mail relay daemon), `tui/internal/timemachine` (git history daemon).
 - **Bootstrap sequence** (`main.go:228-273`): on every launch, the TUI runs `config.MigrateLegacyLanguage` → `config.NeedsVenv` (for setup banner) → `config.EnsureRuntime` (create/repair venv if needed, then always run the non-blocking `CheckUpgrade`) → `config.EnsureAddons` → `preset.Bootstrap` → `tui.ExportCommandsJSON` → `maybePromptRustToolchain` (one-time optional Rust/Cargo prompt only when file search is on Python fallback and no cargo is on PATH). `CheckUpgrade` auto-upgrades the `lingtai` meta-package from PyPI, which bundles `lingtai-kernel` + all addon MCPs. See `tui/internal/config/ANATOMY.md`.
 - **`lingtai-tui doctor` subcommand** (`main.go:962-1002`): runs `config.RunDoctorUpdate` (`main.go:974`) with both `ForceTUI=true` and `ForcePython=true`, then refreshes presets, utility skills, and `commands.json`. The report includes native file-search sidecar / Rust toolchain diagnostics (`config.checkFileSearchNative`). Designed to be usable when the TUI cannot start (broken venv, missing migrations) — it never touches `.lingtai/`. Exit nonzero only on unrecoverable failures.
 - **Version flow:** `Makefile:4` injects `git describe` into `main.go:31`. On startup, `main.go:112` calls `tui.SetTUIVersion(version)`, which stores it for `/doctor` drift detection.
@@ -47,6 +47,7 @@ This folder is the self-contained Go module for the `lingtai-tui` terminal UI bi
   - `tui/internal/fs/` — filesystem readers for agent state
   - `tui/internal/config/` — bootstrap, venv, global config (`tui/internal/config/ANATOMY.md`)
   - `tui/internal/process/` — subprocess launcher
+  - `tui/internal/processscan/` — shared ps-based `lingtai run <agentDir>` process detection used by launch and migrations
   - `tui/internal/headless/` — JSON-emitting non-interactive CLI surface (`bootstrap`, `presets`, `spawn` subcommands)
   - `tui/internal/postman/` — UDP cross-internet mail relay
   - `tui/internal/timemachine/` — git-backed history daemon
